Updated package configuration
- Set minimum supported PHP version to 5.6
- Updated all component and module minimum supported versions to known stable
  versions; if v3 releases are available, supported via `||` constraints.
- Expose module to zend-component-installer
- Move `Module` class to `src/` tree, where it belongs, and:
  - Remove `getAutoloaderConfig()`
  - Remove direct interface implementations
  - Use `::class` notation
