--- Module.php
@@ -1,42 +1,0 @@
-<?php
-/**
- * @license   http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
- * @copyright Copyright (c) 2014 Zend Technologies USA Inc. (http://www.zend.com)
- */
-
-namespace ZF\Apigility\Documentation;
-
-use Zend\ModuleManager\Feature\AutoloaderProviderInterface;
-use Zend\ModuleManager\Feature\ConfigProviderInterface;
-
-class Module implements AutoloaderProviderInterface, ConfigProviderInterface
-{
-    public function getConfig()
-    {
-        return include __DIR__ . '/config/module.config.php';
-    }
-
-    public function getAutoloaderConfig()
-    {
-        return array(
-            'Zend\Loader\StandardAutoloader' => array(
-                'namespaces' => array(
-                    __NAMESPACE__ => __DIR__ . '/src/',
-                ),
-            ),
-        );
-    }
-
-    public function getServiceConfig()
-    {
-        return array('factories' => array(
-            'ZF\Apigility\Documentation\ApiFactory' => function ($services) {
-                    return new ApiFactory(
-                        $services->get('ModuleManager'),
-                        $services->get('Config'),
-                        $services->get('ZF\Configuration\ModuleUtils')
-                    );
-                }
-        ));
-    }
-}
