+++ composer.json
@@ -21,36 +21,44 @@
         "branch-alias": {
             "dev-master": "1.1-dev",
             "dev-develop": "1.2-dev"
+        },
+        "zf": {
+            "module": "ZF\\Apigility\\Documentation"
         }
     },
     "require": {
-        "php": ">=5.5",
+        "php": "^5.6 || ^7.0",
         "michelf/php-markdown" : "^1.5",
-        "zendframework/zend-modulemanager": "~2.3",
-        "zendframework/zend-mvc": "~2.3",
-        "zendframework/zend-servicemanager": "~2.3",
-        "zendframework/zend-view": "~2.3",
-        "zfcampus/zf-apigility-provider": "~1.0",
-        "zfcampus/zf-configuration": "~1.0",
-        "zfcampus/zf-content-negotiation": "~1.0"
+        "zendframework/zend-modulemanager": "^2.7.2",
+        "zendframework/zend-mvc": "^2.7.9 || ^3.0.2",
+        "zendframework/zend-servicemanager": "^2.7.6 || ^3.1",
+        "zendframework/zend-view": "^2.8.1",
+        "zfcampus/zf-apigility-provider": "^1.2",
+        "zfcampus/zf-configuration": "^1.2",
+        "zfcampus/zf-content-negotiation": "^1.2.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.7",
+        "phpunit/phpunit": "^4.8 || ^5.0",
         "squizlabs/php_codesniffer": "^2.3.1",
-        "zendframework/zend-loader": "~2.3",
-        "zendframework/zend-stdlib": "~2.3"
+        "zendframework/zend-stdlib": "^2.7.7 || ^3.0.1"
     },
     "autoload": {
         "psr-4": {
             "ZF\\Apigility\\Documentation\\": "src/"
-        },
-        "classmap": [
-            "Module.php"
-        ]
+        }
     },
     "autoload-dev": {
         "psr-4": {
             "ZFTest\\Apigility\\Documentation\\": "test/"
         }
+    },
+    "scripts": {
+        "check": [
+            "@cs-check",
+            "@test"
+        ],
+        "cs-check": "phpcs",
+        "cs-fix": "phpcbf",
+        "test": "phpunit"
     }
 }
