+++ src/Module.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * @license   http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
+ * @copyright Copyright (c) 2014-2016 Zend Technologies USA Inc. (http://www.zend.com)
+ */
+
+namespace ZF\Apigility\Documentation;
+
+use ZF\Configuration\ModuleUtils;
+
+class Module
+{
+    public function getConfig()
+    {
+        return include __DIR__ . '/../config/module.config.php';
+    }
+
+    public function getServiceConfig()
+    {
+        return array('factories' => array(
+            ApiFactory::class => function ($services) {
+                return new ApiFactory(
+                    $services->get('ModuleManager'),
+                    $services->get('config'),
+                    $services->get(ModuleUtils::class)
+                );
+            }
+        ));
+    }
+}
